pyproject.toml
[tool.poetry]
name = "pareto"
version = "0.1.0"
description = ""
authors = ["Karlson Pfannschmidt <kiudee@mail.upb.de>"]
license = "Apache-2.0"
classifiers = [
        "Development Status :: 2 - Pre-Alpha",
        "Intended Audience :: Developers",
        "License :: OSI Approved :: Apache Software License",
        "Natural Language :: English",
        "Programming Language :: Python :: 3",
        "Programming Language :: Python :: 3.7",
        "Programming Language :: Python :: 3.8"
]

[tool.poetry.dependencies]
python = "^3.7"
bask = ">=0.9,<1.0"
scikit-learn = ">=0.23,<0.25"
numpy = "^1.18.1"
scipy = "^1.4.1"
skorch = "^0.8"
torch = "^1.6"
torchvision = "^0.7"
pygmo = "^2.14.1"
sqlalchemy = {version = "==1.3.15", optional = true}
psycopg2 = {version = "==2.8.4", optional = true}
alembic = {version = "^1.4.2", optional = true}
dill = ">=0.3.2,<1.0.0"
csrank = {git = "https://github.com/kiudee/cs-ranking.git", optional = true, rev = "d05ca1d5202f0cfd6bea0805bb1c20c86853d770"}
tensorflow-gpu = {version = "^1.15.4", optional = true}
gorilla = {version = "^0.3", optional = true}

[tool.poetry.extras]
pgsql = ["psycopg2", "sqlalchemy", "alembic"]
exp = ["csrank", "gorilla", "tensorflow-gpu"]

[tool.poetry.dev-dependencies]
pytest = "^5.4.1"
black = "^19.10b0"
pandas = "^1.0.3"
ipykernel = "^5.3.0"
flake8 = "^3.8.3"
flake8-bugbear = "^20.1.4"
isort = "^5.4.2"
setuptools = "^50.3.2"

[tool.poetry.scripts]
run_experiment = 'pareto.experiments.runner:run_experiment'
fetch_job = 'pareto.experiments.runner:fetch_job'

[tool.isort]
profile = "black"
src_paths = ["pareto", "tests"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
